requirements.txt
yt-dlp
ffmpeg-python 